COMP: Fix pow and specialization.
COMP(Win): Fix warning about conversion from size_t to uint.
Also add explicit parameter init = 0 in ComputeAllIndices,
instead of rely on default parameter.

COMP: Remove typename from non-template Monogenic...Test.

COMP: Solve unused variable in RieszFrequencyFunctionTest.

+++ include/itkRieszRotationMatrix.hxx
@@ -272,7 +272,7 @@ RieszRotationMatrix< T, VImageDimension >
             else
               {
               rotationFactor *= std::pow(
-                this->m_SpatialRotationMatrix[kIndice][dim], k );
+                this->m_SpatialRotationMatrix[kIndice][dim], static_cast<int>(k) );
               }
             }
           }
